Postgres and digital ociean is the vibe i guess
