+++ .env.example
@@ -1,5 +1,2 @@
 PH_API_KEY=
-
-DATABASE_URL="file:./dev.db"
-TURSO_DATABASE_URL=
-TURSO_DATABASE_TOKEN=
+DATABASE_URL=
