+++ src/app/utils/date.ts
@@ -1,0 +1,41 @@
+// Function to get the current date in PST (UTC-8)
+// ProductHunt is in PST
+export function getCurrentDateInPST() {
+  const currentUTCDate = new Date();
+
+  // Get the UTC offset for PST (UTC-8)
+  const pstOffset = 8 * 60 * 60 * 1000;
+
+  // Convert the current date to PST by subtracting the PST offset
+  const pstDate = new Date(currentUTCDate.getTime() - pstOffset);
+
+  // Return the PST date object
+  return pstDate;
+}
+
+// Function to get the start and end of the current day in PST
+// Then convert to UTC ISO-8601 strings
+export function getStartAndEndOfDayInUTC() {
+  const currentPSTDate = getCurrentDateInPST();
+
+  // Get the PST date components
+  const year = currentPSTDate.getUTCFullYear();
+  const month = currentPSTDate.getUTCMonth();
+  const day = currentPSTDate.getUTCDate();
+
+  // Set to the start of the day (12:00 AM) in PST
+  const startOfDayPST = new Date(Date.UTC(year, month, day, 0, 0, 0));
+
+  // Set to the end of the day (11:59 PM) in PST
+  const endOfDayPST = new Date(Date.UTC(year, month, day, 23, 59, 59));
+
+  // Convert PST times to UTC ISO-8601 format
+  const startOfDayUTC = startOfDayPST.toISOString();
+  const endOfDayUTC = endOfDayPST.toISOString();
+
+  // Return the start and end of the day in UTC ISO-8601 format
+  return {
+    postedAfter: startOfDayUTC,
+    postedBefore: endOfDayUTC,
+  };
+}
